Changed planet population counts and trade income fields to double to work better with the continuous money accumulation.

+++ open-ig/src/hu/openig/model/EmpireStatus.java
@@ -25,7 +25,7 @@ public class EmpireStatus implements MessageObjectIO {
 	/** The empire id. */
 	public String id;
 	/** The current money. */
-	public long money;
+	public double money;
 	/** The negotiation offers from players. */
 	public final Map<String, DiplomaticOffer> offers = new LinkedHashMap<>();
 	/** The player level statistics. */
